desloppify: fix logging consistency and split test monolith
Logging:
- Replace all NSLog() calls with VocaLogger in SoundManager, AppState,
  and VocaMacApp for consistent logging across the codebase

Test organization:
- Split VocaMacTests.swift (775 LOC, 17 test classes) into 4 focused files:
  - ModelTests.swift: SystemInfo, ModelSize, ModelManager, WhisperModelInfo,
    TranscriptionResult, AppStatus, ActivationMode
  - ServiceTests.swift: KeyCodeReference, TextInjector, SoundManager, AudioEngine
  - WhisperServiceTests.swift: Translation, Hallucination filtering
  - AppStateTests.swift: Translation toggle, OnboardingStep, LaunchAtLogin

All 136 tests pass.
